fix: Types int for ids changes to int64, float to float64 fix: add foreign keys support for db db:  add client.phone, project.nr, project.address

+++ api/api.go
@@ -23,7 +23,7 @@ var APIErrorDesc = [...]string{
 type Answer struct {
 	Code    APIErrorCode
 	Message string
-	ID      int // id измененного(возвращаемого) элемента
+	ID      int64 // id измененного(возвращаемого) элемента
 	Result  interface{}
 }
